CC | Information
This can be placed in ".properties" files or in "DB" files, the latter option is more secure, and also allows you to encrypt the specified column in MySQL.
Important: AES supports only exact keys of the specified size which can be made up of, I repeat, exactly 16, 24, 32, 128 and up to 256 bytes.
In this file as written, it is 32 bytes. ENCRYPT_KEY (No reference to the user's personal key).